Split new model-building code into a separate file
